Move `bracketed` and `separated` from `expr::explain` to `ore::str`.
