+++ src/dataflow-types/Cargo.toml
@@ -14,6 +14,7 @@ globset = { version = "0.4.0", features = ["serde1"] }
 interchange = { path = "../interchange" }
 kafka-util = { path = "../kafka-util" }
 log = "0.4.13"
+ore = { path = "../ore" }
 regex = "1.4.6"
 repr = { path = "../repr" }
 rusoto_core = { git = "https://github.com/rusoto/rusoto.git" }
