+++ src/expr/src/scalar/mod.rs
@@ -14,6 +14,7 @@ use std::mem;
 use serde::{Deserialize, Serialize};
 
 use ore::collections::CollectionExt;
+use ore::str::separated;
 use repr::adt::array::InvalidArrayError;
 use repr::adt::datetime::DateTimeUnits;
 use repr::adt::regex::Regex;
@@ -869,7 +870,7 @@ impl fmt::Display for MirScalarExpr {
                 }
             }
             CallVariadic { func, exprs } => {
-                write!(f, "{}({})", func, explain::separated(", ", exprs.clone()))?;
+                write!(f, "{}({})", func, separated(", ", exprs.clone()))?;
             }
             If { cond, then, els } => {
                 write!(f, "if {} then {{{}}} else {{{}}}", cond, then, els)?;
