+++ src/sql/src/plan/explain.rs
@@ -21,8 +21,9 @@
 use std::collections::{BTreeMap, HashMap};
 use std::fmt;
 
-use expr::explain::{bracketed, separated, Indices};
+use expr::explain::Indices;
 use expr::{ExprHumanizer, Id, IdGen, RowSetFinishing};
+use ore::str::{bracketed, separated};
 use repr::{RelationType, ScalarType};
 
 use crate::plan::expr::{AggregateExpr, HirRelationExpr, HirScalarExpr};
